+++ accumulative.js
